Refactor dimension views
